pyproject.toml
[build-system]
requires = ["setuptools>=68.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "dsa-agent"
version = "0.1.0"
description = "Data Science Agent: modularized LangChain RAG + EDA assistant"
readme = "README.md"
license = {text = "MIT"}
requires-python = ">=3.11"
authors = [{name = "Steve Archuleta", email = "stevearchuleta@me.com"}]

dependencies = [
    "openai>=1.40",
    "langchain>=0.3",
    "langchain-openai>=0.3",
    "langchain-community>=0.3",
    "langchain-experimental>=0.3",
    "pypdf>=4.0",
    "faiss-cpu>=1.7",
    "tiktoken>=0.7",
    "python-dotenv>=1.0",
    "pandas>=2.1",
    "matplotlib>=3.8",
    "seaborn>=0.13",
]

[project.optional-dependencies]
dev = [
    "ruff>=0.4",
    "pytest>=8.0",
    "pytest-cov>=5.0",
    "ipykernel>=6.29",
]

[tool.setuptools.packages.find]
where = ["src"]

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.pytest.ini_options]
testpaths = ["tests"]
